Add PersonLinkDataMixin
