+++ indico/modules/events/models/persons.py
@@ -19,10 +19,28 @@ from __future__ import unicode_literals
 from sqlalchemy.ext.declarative import declared_attr
 
 from indico.core.db.sqlalchemy import db, PyIntEnum
+from indico.core.db.sqlalchemy.principals import EmailPrincipal
 from indico.core.db.sqlalchemy.util.models import auto_table_args, override_attr
+from indico.core.db.sqlalchemy.util.session import no_autoflush
 from indico.modules.users.models.users import UserTitle, PersonMixin
 from indico.util.decorators import strict_classproperty
 from indico.util.string import return_ascii, format_repr
+
+
+class PersonLinkDataMixin(object):
+    @property
+    def person_link_data(self):
+        return {x: x.is_submitter for x in self.person_links}
+
+    @person_link_data.setter
+    @no_autoflush
+    def person_link_data(self, value):
+        self.person_links = value.keys()
+        for person_link, is_submitter in value.iteritems():
+            person = person_link.person
+            principal = person.user if person.user is not None else EmailPrincipal(person.email)
+            action = {'add_roles': {'submit'}} if is_submitter else {'del_roles': {'submit'}}
+            self.update_principal(principal, **action)
 
 
 class EventPerson(PersonMixin, db.Model):
